Added `variadic` for producing variadic functions
Just makes it easier to deal with variadic functions, by
always making the last param be an array. Not sure about the
way this is implemented. It compiles a function on the fly
using `eval`. Performance tests suggests this is no big deal,
and there should be no security issues since there's no
custom code being compiled.

But it smells. I don't like smelly things. Unless it's
cheese, then it might be ok.