Added `nil`
Will need this after all, since `null` and `undefined` are
perfectly legit values in for instance sequences. A case could
be made that at least `null` should have the semantics of
`nil`, but I think it has too much baggage and is used for too
many other things.

I reckon `nil` is a clean slate.