Improved `variadic` tests and change behavior of `rest`
Previously, `rest` was always defined and passed, even if the
original call didn't supply enough arguments to reach `rest`.
This poses some problems when you expect the length of
`arguments` to match that of the original call to the variadic
function, so now `rest` is only supplied if there are actual
arguments to fill it up with. This means it'll be undefined
otherwise, and `arguments` will be identical to the original
call.

Additionally, I changed `variadic` so that it'll correctly
report `length` for the wrapped function, provided the wrapped
function `length` is 26 or below. I'd be hard pressed to
remember ever seeing a function with more than even 10
parameters, but still. This isn't an ideal solution of course,
but probably hits home 99.99% of the time which is enough.