Added `defprop` for adding properties to an object
This is a wrapper around `Object.defineProperties` which makes
it easier to add immutable and non-enumerable properties to
objects. Just like `Object.defineProperties`, this function
will mutate the object in place.

It *is* possible to make the properties both mutable and
enumerable (but not removable) – however this is by design
made to be uncomfortable.

Accessor functions or values that are functions will have
their scope bound to a private scope, which is an extension of
the target object. This allows functions to share data
internally, without fear of exposing anything to the outside
world. A caveat here is that subsequent calls to `defprop`
will create a *different* private scope, with the assumption
being that such calls imply that the properties are unrelated.
I may want to rethink this at some point.