Changed how property binding works in `defprop`
It was a bit too magical to make up a magical scope, so now it
has to be specified by providing an object.
Added global options for all properties as well, making it a
bit easier to deal with property options. They can still be
overridden for each field however.